+++ changelog.md
@@ -1,11 +1,11 @@
 # Changelog
 
-## Unreleased changes
+## 0.18.7 RawValue 2021-06-18
 
 ### Devices
 
 - Add RawValue device.
-- Remove unused HA-specific attributes (unique_id, device_class, create_sensors)
+- Remove unused HA-specific attributes (unique_id, device_class, create_sensors).
 - Climate: add `group_address_active_state`, `group_address_command_value_state` and a `is_active` property.
 - Configurable `sync_state` in all devices.